re-org files + add server stuff
