--- src/tensorshare/client/__init__.py
@@ -1,8 +1,0 @@
-"""
-The Client module defines the client side of the application. It is responsible
-for sending requests to the server and exchanging data with it.
-"""
-
-from .module import TensorShareClient
-
-__all__ = ["TensorShareClient"]
